Refactor addon into modules
